+++ src/ploomber/sources/notebooksource.py
@@ -40,8 +40,7 @@ from jupytext.formats import long_form_one_format, short_form_one_format
 from jupytext.config import JupytextConfiguration
 import parso
 
-from ploomber.exceptions import (SourceInitializationError,
-                                 MissingParametersCellError)
+from ploomber.exceptions import SourceInitializationError
 from ploomber.placeholders.placeholder import Placeholder
 from ploomber.util import requires
 from ploomber.sources.abc import Source
@@ -318,7 +317,8 @@ class NotebookSource(Source):
         # the latest version
         _, nb = self._read_nb_str_unrendered()
 
-        if 'parameters' in _get_last_cell(nb).metadata.get('tags', []):
+        if nb.cells and 'parameters' in _get_last_cell(nb).metadata.get(
+                'tags', []):
             cell_suggestion = _get_cell_suggestion(nb)
             kind = 'notebook' if self._ext_in == 'ipynb' else 'script'
             raise SourceInitializationError(
@@ -336,7 +336,9 @@ class NotebookSource(Source):
         # section on each cell's metadata, which makes it too verbose when
         # using NotebookRunner.develop() when the source is script (each cell
         # will have an empty "papermill" metadata dictionary)
-        nb = parameterize_notebook(nb, self._params)
+
+        # NOTE: in papermill 2.4.0, this method no longer creates a deepcopy
+        nb = parameterize_notebook(deepcopy(nb), self._params)
 
         # delete empty tags to prevent cluttering the notebooks
         for cell in nb.cells:
@@ -393,34 +395,31 @@ class NotebookSource(Source):
         Validate notebook after initialization (run pyflakes to detect
         syntax errors)
         """
-        # NOTE: what happens if I pass source code with errors to parso?
-        # maybe we don't need to use pyflakes after all
-        # we can also use compile. can pyflakes detect things that
-        # compile cannot?
+        pass
+
+    def _validate_parameters_cell(self,
+                                  extract_upstream=False,
+                                  extract_product=False):
+        '''Check parameters call and add it when it's missing
+
+        Parameters
+        ----------
+        extract_upstream : bool, default: False
+            Flags used to determine the content of the parameters cell,
+            only used if the notebook is missing the parameters cell
+        extract_product : bool, default: False
+            Same as extract_upstream
+        '''
         params_cell, _ = find_cell_with_tag(self._nb_obj_unrendered,
                                             'parameters')
 
         if params_cell is None:
-            loc = ' "{}"'.format(self.loc) if self.loc else ''
-            msg = ('Notebook{} does not have a cell tagged '
-                   '"parameters"'.format(loc))
-
-            if self.loc and Path(self.loc).suffix == '.py':
-                msg += """.
-Add a cell at the top like this:
-
-# %% tags=["parameters"]
-upstream = None
-product = None
-
-Go to: https://ploomber.io/s/params for more information
-"""
-            if self.loc and Path(self.loc).suffix == '.ipynb':
-                msg += ('. Add a cell at the top and tag it as "parameters". '
-                        'Go to the next URL for '
-                        'details: https://ploomber.io/s/params')
-
-            raise MissingParametersCellError(msg)
+            loc = pretty_print.try_relative_path(self.loc)
+            add_parameters_cell(self.loc, extract_upstream, extract_product)
+            click.secho(
+                f'Notebook {loc} is missing the parameters cell, '
+                'adding it at the top of the file...',
+                fg='yellow')
 
     def _post_render_validation(self):
         """
@@ -542,7 +541,7 @@ Go to: https://ploomber.io/s/params for more information
     def _get_parameters_cell(self, force=False):
         self._read_nb_str_unrendered(force=force)
         cell, _ = find_cell_with_tag(self._nb_obj_unrendered, tag='parameters')
-        return cell.source
+        return cell.source if cell else None
 
     def extract_upstream(self, force=False):
         extractor_class = extractor_class_for_language(self.language)
@@ -1019,6 +1018,11 @@ def _nb2codestr(nb):
 def _warn_on_unused_params(nb, params):
     nb = deepcopy(nb)
     _, idx = find_cell_with_tag(nb, 'parameters')
+
+    # the notebooks might not have a parameters cell
+    if idx is None:
+        return
+
     del nb.cells[idx]
 
     code = _nb2codestr(nb)
@@ -1036,7 +1040,7 @@ def _warn_on_unused_params(nb, params):
                       f'task\'s source code: {pretty_print.iterable(unused)}')
 
 
-def add_parameters_cell(path, extract_upstream, extract_product):
+def add_parameters_cell(path, extract_upstream=False, extract_product=False):
     """
     Add parameters cell to a script/notebook in the given path, overwrites the
     original file
@@ -1053,6 +1057,9 @@ upstream = None
 # declare a dictionary with the outputs of this task
 product = None
 """
+
+    if not extract_upstream and not extract_product:
+        source += '# add default values for parameters here'
 
     c = JupytextConfiguration()
     c.notebook_metadata_filter